+++ app/javascript/plugins/init_mapbox.js
@@ -12,11 +12,20 @@ const initMapbox = () => {
     });
     const markers = JSON.parse(mapElement.dataset.markers)
     markers.forEach((marker) => {
+
+      const element = document.createElement('div');
+      element.className = 'marker';
+      element.style.backgroundImage = `url('${marker.image_url}')`;
+      element.style.backgroundSize = 'contian';
+      element.style.width = '45px';
+      element.style.height = '45px';
+
       new mapboxgl.Marker()
       .setLnglat([marker.lng, marker.lat])
       .addTo(map);
     });
     fitMapToMarkers(map,markers)
+    map.addControl(new Mapbox({accessToken: mapboxgl.accessToken,mapboxgl: mapboxgl}));
   }
 };
